🔊 Wrapping most platform calls to log erros
DatadogSdk now has an InternalLogger class that it passes to internal plugins.
Most platform calls are wrapped with `wrap` which catches all exceptions and specifically warns users on ArgumentErrors (which are most likely programmer errors). All other exceptions are caught and logged

Telemetry back to Datadog still needs to be added and those areas are marked with // TELEMETRY