🧹 Cleanup resource name on traces, add IGNORE for e2e monitors
In order to do trace measurement, we need a "resource name" which iOS and Android do slightly differently. Added a `resourceName` parameter to traces to support handling this cross platform.

Since performance on Android is outside where we want it so we're ignoring performance metrics at the moment, until we can push to the plugin without awaiting future. Added IGNORE as a specifier for monitors to support this