🔊 Adding logging messages to Mock Http server, add new example error
When running in simulators, writing sessions puts the files in a weird location. Printing to console is easier in these situations.
Also added a new Flutter error that is non-async, which results in a larger stack that goes further into Flutter's built-in libraries.

Also, added an extra regex test for firstPartyHosts.