✨ Send telemetry to Datadog on errors
Also add a configuration option for the `telemetrySampleRate`.

The `wrap` function does most of the heavy lifting, sending errors to Datadog if any exception occurs other than those that are caused by serialization (which should be reported to the user).

Note, changes to the http_client will need to be brought into 1.0 branch.