🐛 Fix an issue backing out of an Android application
On Android, if you press the back button on the top of the view stack, the app backgrounds and detaches the Flutter engine. We would then clean up a lot of the Flutter -> Datadog glue. However, if the application was then reopened, Datadog would be in a weird state where native code was initialized and thus not re-setup any of the glue.

This changes it so the glue is always set up for all Datadog modules properly, and still cleaned up on a Flutter engine detach.