✨ Add better support to NavigationObserver for AppState scenarios
Changes to NativationObserver are mostly to support Add-To-App scenarios, but it also fixes issues where views were not stopped when the app was backgrounded / foregrounded.
Use pre-cached engines in the add-to-app example and remove documentation warning against pre-cached FlutterEngines,
Add examples of View predicates for iOS and Android to filter out "FlutterViewController" and "FlutterActivity"
Add documentation around View predicates and restarting ViewController views on dismissing FlutterViewControllers.