🐛 Fix an iOS crash on app termination
If the Datadog internal logger attempted to send a console log while the app was terminating, you would get an assert complaining that the engine had not been run yet. This fixes the issue by restoring `consoleLog` to its initial state during engine detach or app termination.

If we have lots of these callbacks, we should potentially switch the method channel to optional, set it to nil during shutdown, and use optional chaining to avoid the crash.

Fixes #271