💚 Fix flakiness with new HttpClient interceptor tests
Tracking long tasks was causing a lot of traffic in certain situations, causing the http capture to slow down.
Also, because the last view doesn't do anything, it looks like android specifically was delaying sending it. Since we don't need it, only look for at least 3 views in the test.