🐛 Allow sdkVerbosity to be set prior to WidgetsFlutterBinding being initialized
Because we currently use runZonedGuarded to catch errors, we need to make sure WidgetsFlutterBinding.ensureInitialized is called inside runZonedGuarded.

This is an issue if you try to set sdkVerbosity (which requires the binding be initialized) before calling DatadogSdk.runApp. This fixes the issue by not calling into the platform channel until Datadog is initialized or during the initialization process.