🐛 Fix rethrown exceptions not properly preserving stacktrace
Dart has "rethrow" within catch blocks to preserve stack traces, and Futures will preserve stack traces if you "throw" the same exception so long as it is not inside an `async` block. The `close` method had its `onError` marked as async by accident, so it was not preserving the stack trace properly.

Fixes RUMM-3264