🐛 Attempt to fix crash on exit on iOS
Another attempt to fix Flutter shutting down its engine without telling us.  In this case, we're going to slightly delay sending the message by jumping to the main thread, and check to make sure the FlutterEngine is working before we send off the console log.

In the case that we cannot find the FlutterViewController at the root, we'll bypass the engine check, as that will be case for add-to-app scenarios, and I'm hoping the jump to main thread will solve the issues for those apps.

Fixes #414

(cherry picked from commit 9aceee5df37372a2e5fe2ba83e2e6658fc94d97b)