fix: Loggers incorrectly used themselves as their finalizer token
Objects in Dart cannot use themselves as finalizer tokens, otherwise they create a circular reference that leaks. I made this mistake in the initial implementation, but new versions of Dart throw if you do that.

This uses the logger handle as the finalization token (which was the only thing we needed anyway).

Also add the missing changelog item, and remove the no longer needed CI configuration.