fix: Move `LogLevel` to public file
LogLevel is now used for `DatadogLogConfiguration.remoteLogLevel`, whereas before it was internal. Move it to a public file so that the remote log level can be configured.

LogLevel currently has all possible log levels, though only some are supported. Unsupported levels are documented, but it shouldn't be possible to send a log with an unsupported level.

Also update some out of date documentation.

refs: #518