fix:  Fix Android losing data on backgrounding
Android is recreating the main DatadogSdkPlugin on background / foregrounding the application, which recreates the DatadogRumPlugin, which loses its previous reference to the RUMMonitor.  This is causing telemetry to report that RUM is being used when it's not enabled, and likely leading to lost data.

The fast fix is to have DatadogRumPlugin be a singleton, which is what's done here. I've made DatadogLogsPlugin one as well, although that was likely not needed.