fix(ios): Don't force unwrap `currentConfiguration`
In the instance where `initialize` is called but Datadog is already initialized, if no previous configuration existed the force unwrap would cause a crash.  This can happen in cases where multiple calls to `initialize` were made in multiple engines.

This does not fix multiple engines support, but prevents the crash.