fix: Error setting firstPartyHosts in Flutter Web
The Browser SDK is checking if the type supplied to firstPartyHosts is a RegExp, but the Dart RegExp returns `false` for that check.  Fix it so that we create a JavaScript RegExp.

Note, this only works because the RegExp is very simple in this case. More complicated RegExes may not convert properly.

refs: #554