chore: Remove deprecated `apply plugin` from gradle files
Flutter has deprecated using `apply plugin` for its gradle plugin, and it issues a warning if you try to use it.  This cleans up the use of `apply plugin` from all of our example and test projects.

Additionally, switch the Dart language version to "<4.0.0", as that is how Dart is interpreting it anyway (and publishing it as is right now is issuing a warning).

refs: RUM-3500