fix: Fix exception when encoding variables in gql_link
The gel_link encodes variables to JSON for display in the Datadog UI. However, there are situations when those variables are not encodable (such as when sending a file using `MultipartFile`.  In these situations, we'll take the `toString` of the variable instead.

To avoid this issue in the future, wrap the `jsonEncode` in a try / catch and send encoding issues to both the developer and to telemetry.

refs: #618