fix: Prevent error when using logging from background isolate
When creating a logger from a background isolate, Flutter would throw because WidgetsFlutterBinding wasn't initialized. This is specifically because we were adding a callback handler, which is not supported from background isolates.

Add checks to both logs and RUM and prevent initialization of things we know won't work from background isolates.

This is not full support for background isolates but usable as a work around.

refs: #580