fix: Wait 1ms during route changes on web
The Browser SDK reads `location` during a call to `startView`. Unfortunately, the RouterObserver is called before the browser's location has changed, leading to a disconnect in what we send in the event.

Wait 1ms after a route change and, if the route hasn't changed in that 1ms, then call startView.

refs: RUM-5416