fix: Add additional timeout for mappers
Additionally, attempt to call other mappers if the first one returns NotImplemented. This is to avoid issues with multiple Flutter engines or some weird (and very difficult to reproduce cases) where it appears that Flutter does not unregister the plugin.

Add formerly missing mapper "not implemented" telemetry to iOS.

Add an integration test for user information.

refs: RUM-6457 #652